feat: hide sidebar on unmatched routes
